Updated model constants (primarily loss years/model years) for annual update. Also tentatively updated input and output folder dates on s3 from 2021 to 2022. Finally, refactored "gain" folder to "removals" folder (don't know if refactor worked entirely correctly but it did results in "gain" being changed to "removals" in comments and variables throughout the model).
