+++ emissions/cpp_util/constants.h
@@ -5,7 +5,7 @@ namespace constants
 {
     // Emissions constants
     // per https://www.learncpp.com/cpp-tutorial/global-constants-and-inline-variables/
-    constexpr int model_years {20};    // How many loss years are in the model. Must also be updated in equations.cpp!
+    constexpr int model_years {21};    // How many loss years are in the model. Must also be updated in equations.cpp!
 
     constexpr int CH4_equiv {28};      // The CO2 equivalency (global warming potential) of CH4
 
@@ -34,7 +34,7 @@ namespace constants
 
     constexpr char legal_Amazon_loss[] = "_legal_Amazon_annual_loss_2001_2019.tif";
 
-    constexpr char lossyear[] = "GFW2020_";
+    constexpr char lossyear[] = "GFW2021_";
     constexpr char burnyear[] = "_burnyear_with_Hansen_loss.tif";
     constexpr char fao_ecozones[] = "_fao_ecozones_bor_tem_tro_processed.tif";
     constexpr char climate_zones[] = "_climate_zone_processed.tif";
