+++ emissions/cpp_util/equations.cpp
@@ -17,7 +17,7 @@ void def_variables(float *q, int ecozone, int forestmodel_data, int ifl, int cli
 {
 
 	int model_years;    // How many loss years are in the model
-    model_years = 20;
+    model_years = 21;
 
 	int tropical;       // The ecozone code for the tropics
     tropical = 1;
