+++ removals/.gitignore
