+++ removals/annual_gain_rate_AGC_BGC_all_forest_types.py
