+++ removals/mp_US_removal_rates.py
@@ -103,7 +103,7 @@ def mp_US_removal_rates(sensit_type, tile_id_list, run_date):
     gain_table = pd.read_excel("{}".format(cn.table_US_removal_rate),
                                sheet_name="US_rates_AGC+BGC")
 
-    # Converts gain table from wide to long, so each region-group-age category has its own row
+    # Converts removals table from wide to long, so each region-group-age category has its own row
     gain_table_group_region_by_age = pd.melt(gain_table, id_vars=['FIA_region_code', 'forest_group_code'],
                                       value_vars=['growth_young', 'growth_middle',
                                                   'growth_old'])
@@ -113,14 +113,14 @@ def mp_US_removal_rates(sensit_type, tile_id_list, run_date):
     age_dict = {'growth_young': 1000, 'growth_middle': 2000, 'growth_old': 3000}
 
     # Creates a unique value for each forest group-region-age category in the table.
-    # Although these rates are applied to all standard gain model pixels at first, they are not ultimately used for
+    # Although these rates are applied to all standard removals model pixels at first, they are not ultimately used for
     # pixels that have Hansen gain (see below).
     gain_table_group_region_age = gain_table_group_region_by_age.replace({"variable": age_dict})
     gain_table_group_region_age['age_cat'] = gain_table_group_region_age['variable']*10
     gain_table_group_region_age['group_region_age_combined'] = gain_table_group_region_age['age_cat'] + \
                                               gain_table_group_region_age['forest_group_code']*100 + \
                                               gain_table_group_region_age['FIA_region_code']
-    # Converts the forest group-region-age codes and corresponding gain rates to a dictionary,
+    # Converts the forest group-region-age codes and corresponding removals rates to a dictionary,
     # where the key is the unique group-region-age code and the value is the AGB removal rate.
     gain_table_group_region_age_dict = pd.Series(gain_table_group_region_age.value.values, index=gain_table_group_region_age.group_region_age_combined).to_dict()
     uu.print_log(gain_table_group_region_age_dict)
@@ -132,7 +132,7 @@ def mp_US_removal_rates(sensit_type, tile_id_list, run_date):
     gain_table_group_region = gain_table_group_region_age.drop(gain_table_group_region_age[gain_table_group_region_age.age_cat != 10000].index)
     gain_table_group_region['group_region_combined'] = gain_table_group_region['forest_group_code']*100 + \
                                                        gain_table_group_region['FIA_region_code']
-    # Converts the forest group-region codes and corresponding gain rates to a dictionary,
+    # Converts the forest group-region codes and corresponding removals rates to a dictionary,
     # where the key is the unique group-region code (youngest age category) and the value is the AGB removal rate.
     gain_table_group_region_dict = pd.Series(gain_table_group_region.value.values, index=gain_table_group_region.group_region_combined).to_dict()
     uu.print_log(gain_table_group_region_dict)
@@ -140,7 +140,7 @@ def mp_US_removal_rates(sensit_type, tile_id_list, run_date):
 
     ### To make the removal factor standard deviation dictionaries
 
-    # Converts gain table from wide to long, so each region-group-age category has its own row
+    # Converts removals table from wide to long, so each region-group-age category has its own row
     stdev_table_group_region_by_age = pd.melt(gain_table, id_vars=['FIA_region_code', 'forest_group_code'],
                                              value_vars=['SD_young', 'SD_middle',
                                                          'SD_old'])
@@ -150,14 +150,14 @@ def mp_US_removal_rates(sensit_type, tile_id_list, run_date):
     stdev_dict = {'SD_young': 1000, 'SD_middle': 2000, 'SD_old': 3000}
 
     # Creates a unique value for each forest group-region-age category in the table.
-    # Although these rates are applied to all standard gain model pixels at first, they are not ultimately used for
+    # Although these rates are applied to all standard removals model pixels at first, they are not ultimately used for
     # pixels that have Hansen gain (see below).
     stdev_table_group_region_age = stdev_table_group_region_by_age.replace({"variable": stdev_dict})
     stdev_table_group_region_age['age_cat'] = stdev_table_group_region_age['variable'] * 10
     stdev_table_group_region_age['group_region_age_combined'] = stdev_table_group_region_age['age_cat'] + \
                                                                stdev_table_group_region_age['forest_group_code'] * 100 + \
                                                                stdev_table_group_region_age['FIA_region_code']
-    # Converts the forest group-region-age codes and corresponding gain rates to a dictionary,
+    # Converts the forest group-region-age codes and corresponding removals rates to a dictionary,
     # where the key is the unique group-region-age code and the value is the AGB removal rate.
     stdev_table_group_region_age_dict = pd.Series(stdev_table_group_region_age.value.values,
                                                  index=stdev_table_group_region_age.group_region_age_combined).to_dict()
@@ -170,7 +170,7 @@ def mp_US_removal_rates(sensit_type, tile_id_list, run_date):
         stdev_table_group_region_age[stdev_table_group_region_age.age_cat != 10000].index)
     stdev_table_group_region['group_region_combined'] = stdev_table_group_region['forest_group_code'] * 100 + \
                                                        stdev_table_group_region['FIA_region_code']
-    # Converts the forest group-region codes and corresponding gain rates to a dictionary,
+    # Converts the forest group-region codes and corresponding removals rates to a dictionary,
     # where the key is the unique group-region code (youngest age category) and the value is the AGB removal rate.
     stdev_table_group_region_dict = pd.Series(stdev_table_group_region.value.values,
                                              index=stdev_table_group_region.group_region_combined).to_dict()
