+++ removals/mp_annual_gain_rate_mangrove.py
@@ -1,7 +1,7 @@
 '''
 Creates tiles of annual aboveground and belowground biomass removal rates for mangroves using IPCC Wetlands Supplement Table 4.4 rates.
 Its inputs are the continent-ecozone tiles, mangrove biomass tiles (for locations of mangroves), and the IPCC mangrove
-gain rate table.
+removals rate table.
 Also creates tiles of standard deviation in mangrove aboveground biomass removal rates based on the 95% CI in IPCC Wetlands Supplement Table 4.4.
 '''
 
@@ -16,7 +16,7 @@ import sys
 sys.path.append('../')
 import constants_and_names as cn
 import universal_util as uu
-sys.path.append(os.path.join(cn.docker_app,'gain'))
+sys.path.append(os.path.join(cn.docker_app,'removals'))
 import annual_gain_rate_mangrove
 
 def mp_annual_gain_rate_mangrove(sensit_type, tile_id_list, run_date = None):
@@ -28,7 +28,7 @@ def mp_annual_gain_rate_mangrove(sensit_type, tile_id_list, run_date = None):
     # If a full model run is specified, the correct set of tiles for the particular script is listed
     if tile_id_list == 'all':
         # Lists the tiles that have both mangrove biomass and FAO ecozone information because both of these are necessary for
-        # calculating mangrove gain
+        # calculating mangrove removals
         mangrove_biomass_tile_list = uu.tile_list_s3(cn.mangrove_biomass_2000_dir)
         ecozone_tile_list = uu.tile_list_s3(cn.cont_eco_dir)
         tile_id_list = list(set(mangrove_biomass_tile_list).intersection(ecozone_tile_list))
@@ -60,14 +60,14 @@ def mp_annual_gain_rate_mangrove(sensit_type, tile_id_list, run_date = None):
         uu.s3_flexible_download(dir, pattern, cn.docker_base_dir, sensit_type, tile_id_list)
 
 
-    # Table with IPCC Wetland Supplement Table 4.4 default mangrove gain rates
+    # Table with IPCC Wetland Supplement Table 4.4 default mangrove removals rates
     cmd = ['aws', 's3', 'cp', os.path.join(cn.gain_spreadsheet_dir, cn.gain_spreadsheet), cn.docker_base_dir, '--no-sign-request']
     uu.log_subprocess_output_full(cmd)
 
 
     ### To make the removal factor dictionaries
 
-    # Imports the table with the ecozone-continent codes and the carbon gain rates
+    # Imports the table with the ecozone-continent codes and the carbon removals rates
     gain_table = pd.read_excel("{}".format(cn.gain_spreadsheet),
                                sheet_name = "mangrove gain, for model")
 
@@ -75,18 +75,18 @@ def mp_annual_gain_rate_mangrove(sensit_type, tile_id_list, run_date = None):
     gain_table_simplified = gain_table.drop_duplicates(subset='gainEcoCon', keep='first')
 
     # Creates belowground:aboveground biomass ratio dictionary for the three mangrove types, where the keys correspond to
-    # the "mangType" field in the gain rate spreadsheet.
+    # the "mangType" field in the removals rate spreadsheet.
     # If the assignment of mangTypes to ecozones changes, that column in the spreadsheet may need to change and the
     # keys in this dictionary would need to change accordingly.
     type_ratio_dict = {'1': cn.below_to_above_trop_dry_mang, '2'  :cn.below_to_above_trop_wet_mang, '3': cn.below_to_above_subtrop_mang}
     type_ratio_dict_final = {int(k):float(v) for k,v in list(type_ratio_dict.items())}
 
-    # Applies the belowground:aboveground biomass ratios for the three mangrove types to the annual aboveground gain rates to get
-    # a column of belowground annual gain rates by mangrove type
+    # Applies the belowground:aboveground biomass ratios for the three mangrove types to the annual aboveground removals rates to get
+    # a column of belowground annual removals rates by mangrove type
     gain_table_simplified['BGB_AGB_ratio'] = gain_table_simplified['mangType'].map(type_ratio_dict_final)
     gain_table_simplified['BGB_annual_rate'] = gain_table_simplified.AGB_gain_tons_ha_yr * gain_table_simplified.BGB_AGB_ratio
 
-    # Converts the continent-ecozone codes and corresponding gain rates to dictionaries for aboveground and belowground gain rates
+    # Converts the continent-ecozone codes and corresponding removals rates to dictionaries for aboveground and belowground removals rates
     gain_above_dict = pd.Series(gain_table_simplified.AGB_gain_tons_ha_yr.values,index=gain_table_simplified.gainEcoCon).to_dict()
     gain_below_dict = pd.Series(gain_table_simplified.BGB_annual_rate.values,index=gain_table_simplified.gainEcoCon).to_dict()
 
@@ -101,14 +101,14 @@ def mp_annual_gain_rate_mangrove(sensit_type, tile_id_list, run_date = None):
 
     ### To make the removal factor standard deviation dictionary
 
-    # Imports the table with the ecozone-continent codes and the carbon gain rates
+    # Imports the table with the ecozone-continent codes and the carbon removals rates
     stdev_table = pd.read_excel("{}".format(cn.gain_spreadsheet),
                                sheet_name = "mangrove stdev, for model")
 
     # Removes rows with duplicate codes (N. and S. America for the same ecozone)
     stdev_table_simplified = stdev_table.drop_duplicates(subset='gainEcoCon', keep='first')
 
-    # Converts the continent-ecozone codes and corresponding gain rate standard deviations to dictionaries for aboveground and belowground gain rate stdevs
+    # Converts the continent-ecozone codes and corresponding removals rate standard deviations to dictionaries for aboveground and belowground removals rate stdevs
     stdev_dict = pd.Series(stdev_table_simplified.AGB_gain_stdev_tons_ha_yr.values,index=stdev_table_simplified.gainEcoCon).to_dict()
 
     # Adds a dictionary entry for where the ecozone-continent code is 0 (not in a continent)
@@ -125,7 +125,7 @@ def mp_annual_gain_rate_mangrove(sensit_type, tile_id_list, run_date = None):
         processes = 20    #26 processors = >740 GB peak; 18 = 550 GB peak; 20 = 610 GB peak; 23 = 700 GB peak; 24 > 750 GB peak
     else:
         processes = 4
-    uu.print_log('Mangrove annual gain rate max processors=', processes)
+    uu.print_log('Mangrove annual removals rate max processors=', processes)
     pool = multiprocessing.Pool(processes)
     pool.map(partial(annual_gain_rate_mangrove.annual_gain_rate, sensit_type=sensit_type, output_pattern_list=output_pattern_list,
                      gain_above_dict=gain_above_dict, gain_below_dict=gain_below_dict, stdev_dict=stdev_dict), tile_id_list)
