+++ removals/mp_gain_year_count_all_forest_types.py
